feat: navigate to replaced file
