+++ src/sections/shared/file-uploader-panel/FileUploader.tsx
@@ -8,7 +8,6 @@ import {
   useRef,
   useState
 } from 'react'
-import { useBlocker } from 'react-router-dom'
 import { useDeepCompareEffect } from 'use-deep-compare'
 import { Trans, useTranslation } from 'react-i18next'
 import { ExclamationTriangle, Plus, XLg } from 'react-bootstrap-icons'
@@ -25,7 +24,6 @@ import { FileUploadState, useFileUploader } from './fileUploaderReducer'
 import { FileUploaderHelper } from './FileUploaderHelper'
 import { SwalModal } from '../swal-modal/SwalModal'
 import { LoadingConfigSpinner } from './loading-config-spinner/LoadingConfigSpinner'
-import { ConfirmLeaveModal } from './confirm-leave-modal/ConfirmLeaveModal'
 import styles from './FileUploader.module.scss'
 
 type FileUploaderProps = {
@@ -78,9 +76,6 @@ const FileUploader = forwardRef<FileUploaderRef, FileUploaderProps>(
       (file) => file.done && file.checksumValue
     )
     const canKeepUploading = multiple ? true : totalFiles === 0
-
-    // Block navigation if there are files in progress or already uploaded
-    const navigationBlocker = useBlocker(totalFiles > 0)
 
     const onFileUploadFailed = (file: File) => {
       setUploadingToCancelMap((x) => {
@@ -272,28 +267,6 @@ const FileUploader = forwardRef<FileUploaderRef, FileUploaderProps>(
       return result.isConfirmed
     }
 
-    const handleConfirmLeavePage = () => {
-      if (navigationBlocker.state === 'blocked') {
-        // First cancel all uploads in progress
-        uploadingToCancelMap.forEach((cancelFunction) => {
-          cancelFunction()
-        })
-
-        // TODO - Remove already uploaded files from the bucket, we need an endpoint for this
-
-        // Then reset the file uploader state
-        removeAllFiles()
-
-        navigationBlocker.proceed()
-      }
-    }
-
-    const handleCancelLeavePage = () => {
-      if (navigationBlocker.state === 'blocked') {
-        navigationBlocker.reset()
-      }
-    }
-
     useImperativeHandle(ref, () => ({
       removeUploadedFile: (fileKey: string) => {
         removeFile(fileKey)
@@ -309,6 +282,8 @@ const FileUploader = forwardRef<FileUploaderRef, FileUploaderProps>(
         setConfig({ checksumAlgorithm: fixityAlgorithm })
       }
     }, [fixityAlgorithm, isLoadingFixityAlgorithm, setConfig])
+
+    // TODO:ME - UseEffect here with return clean function that cancels all uploads in progress
 
     if (isLoadingFixityAlgorithm) return <LoadingConfigSpinner />
 
@@ -423,12 +398,6 @@ const FileUploader = forwardRef<FileUploaderRef, FileUploaderProps>(
             </Accordion.Body>
           </Accordion.Item>
         </Accordion>
-
-        <ConfirmLeaveModal
-          show={navigationBlocker.state === 'blocked'}
-          handleCancelLeavePage={handleCancelLeavePage}
-          handleConfirmLeavePage={handleConfirmLeavePage}
-        />
       </div>
     )
   }
