+++ src/sections/shared/file-uploader-panel/FileUploaderPanel.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useRef, useState } from 'react'
+import { useCallback, useEffect, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Tabs } from '@iqss/dataverse-design-system'
 import { FileRepository } from '@/files/domain/repositories/FileRepository'
@@ -8,6 +8,8 @@ import { FilesListFormData, UploadedFilesList } from './uploaded-files-list/Uplo
 import FileUploader, { FileUploaderRef } from './FileUploader'
 import { FileUploadState } from './fileUploaderReducer'
 import styles from './FileUploaderPanel.module.scss'
+import { useBlocker } from 'react-router-dom'
+import { ConfirmLeaveModal } from './confirm-leave-modal/ConfirmLeaveModal'
 
 type FileUploaderPanelProps =
   | {
@@ -18,6 +20,7 @@ type FileUploaderPanelProps =
       replaceFile?: false
       originalFileType?: never
       onSaveChanges: (data: FilesListFormData) => Promise<void>
+      saveSucceeded: boolean
     }
   | {
       fileRepository: FileRepository
@@ -27,6 +30,7 @@ type FileUploaderPanelProps =
       replaceFile: true
       originalFileType: string
       onSaveChanges: (data: FilesListFormData) => Promise<void>
+      saveSucceeded: boolean
     }
 
 type FileStorageConfiguration = 'S3'
@@ -38,7 +42,8 @@ export const FileUploaderPanel = ({
   multiple,
   replaceFile,
   originalFileType,
-  onSaveChanges
+  onSaveChanges,
+  saveSucceeded
 }: FileUploaderPanelProps) => {
   const { t: tFiles } = useTranslation('files')
   const fileUploaderRef = useRef<FileUploaderRef>(null)
@@ -46,6 +51,22 @@ export const FileUploaderPanel = ({
     []
     // UploadedFilesListHelper.mapUploadedFilesToUploadedFileInfo(Object.values(mockFileUploadState))
   )
+
+  // Block navigation if there are files in progress or already uploaded
+  const navigationBlocker = useBlocker(uploadedFilesInfo.length > 0)
+
+  const handleConfirmLeavePage = () => {
+    if (navigationBlocker.state === 'blocked') {
+      // TODO - Remove already uploaded files from the bucket, we need an endpoint for this
+      navigationBlocker.proceed()
+    }
+  }
+
+  const handleCancelLeavePage = () => {
+    if (navigationBlocker.state === 'blocked') {
+      navigationBlocker.reset()
+    }
+  }
 
   const handleSyncUploadedFiles = useCallback((files: FileUploadState[]) => {
     const uploadedFilesMapped = UploadedFilesListHelper.mapUploadedFilesToUploadedFileInfo(files)
@@ -55,6 +76,17 @@ export const FileUploaderPanel = ({
   const handleRemoveFileFromFileUploaderState = (fileKey: string) => {
     fileUploaderRef.current?.removeUploadedFile(fileKey)
   }
+
+  // After saving successfully, remove the uploaded files from the state and unblock navigation
+  useEffect(() => {
+    if (saveSucceeded) {
+      setUploadedFilesInfo([])
+      if (navigationBlocker.state === 'blocked') {
+        navigationBlocker.proceed()
+      }
+    }
+  }, [saveSucceeded, navigationBlocker])
+
   return (
     <>
       <Tabs defaultActiveKey="metadata">
@@ -81,6 +113,11 @@ export const FileUploaderPanel = ({
           removeFileFromFileUploaderState={handleRemoveFileFromFileUploaderState}
         />
       )}
+      <ConfirmLeaveModal
+        show={navigationBlocker.state === 'blocked'}
+        handleCancelLeavePage={handleCancelLeavePage}
+        handleConfirmLeavePage={handleConfirmLeavePage}
+      />
     </>
   )
 }
